Implement GeneratePressControl to reduce code duplication
