+++ src/customizer-controls/font-manager/index.js
@@ -1,5 +1,10 @@
-import GeneratePressFontManagerControl from './GeneratePressFontManagerControl';
-import GeneratePressTypographyControl from './GeneratePressTypographyControl';
+import GeneratePressControl from "../../components/GeneratePressControl";
+
+import GeneratePressFontManagerControlForm from "./GeneratePressFontManagerControlForm";
+import GeneratePressTypographyControlForm from "./GeneratePressTypographyControlForm";
+
+const GeneratePressFontManagerControl = GeneratePressControl.extend( GeneratePressFontManagerControlForm );
+const GeneratePressTypographyControl = GeneratePressControl.extend( GeneratePressTypographyControlForm );
 
 // Register control type with Customizer.
 wp.customize.controlConstructor[ 'generate-font-manager-control' ] = GeneratePressFontManagerControl;
